Update 수학은 체육과목 입니다.java

+++ 백준/Bronze/15894. 수학은 체육과목 입니다/수학은 체육과목 입니다.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        long N = sc.nextInt();
+        long N = sc.nextLong();
 
         System.out.println(N*4);
     }
